refactor test for users
